breaking out the file service into multiple files, adding configuration
