+++ Cargo.toml
@@ -15,3 +15,5 @@ reqwest = {version = "0.11.22", features=["stream"]}
 tokio = { version = "1", features = ["full"] }
 tokio-util = "0.7.9"
 time = "0.3.31"
+openssl = { version = "0.10", features = ["vendored"] }
+config = "0.13.1"
