+++ daemon/Settings.toml
@@ -1,0 +1,4 @@
+level = "info"
+base_url = "http://localhost:9100"
+data_dir = "./data"
+sleep_interval = 3600
