+++ parquet_file_service/Cargo.toml
@@ -17,6 +17,15 @@ mime = "0.3.17"
 tower-http = { version= "0.4.4", features = ["fs", "cors"] }
 futures = "0.3.28"
 serde = { version="1.0.188", features= ["derive"]}
+openssl = { version = "0.10", features = ["vendored"] }
+num_cpus = "1.16.0"
+config = "0.13.1"
+log = "0.4.18"
+slog = "2.7.0"
+slog-term = "2.9.0"
+slog-async = "2.7.0"
+clap = { version = "4.3.23", features = ["derive", "env"] }
+
 
 [dev-dependencies]
 tower = "0.4.13"
