+++ parquet_file_service/Settings.toml
@@ -1,0 +1,4 @@
+level = "info"
+data_dir = "./weather_data"
+host = "127.0.0.1"
+port = 9100
