+++ parquet_file_service/src/lib.rs
@@ -1,0 +1,7 @@
+mod utils;
+mod routes;
+mod startup;
+
+pub use utils::*;
+pub use routes::*;
+pub use startup::*;
