+++ parquet_file_service/src/routes/files/mod.rs
@@ -1,0 +1,7 @@
+mod get;
+mod get_names;
+mod post;
+
+pub use get::*;
+pub use get_names::*;
+pub use post::*;
