+++ parquet_file_service/src/routes/home/mod.rs
@@ -1,0 +1,3 @@
+mod get;
+
+pub use get::*;
