+++ parquet_file_service/src/routes/mod.rs
@@ -1,0 +1,5 @@
+mod files;
+mod home;
+
+pub use files::*;
+pub use home::*;
