+++ parquet_file_service/src/utils.rs
@@ -1,0 +1,58 @@
+use std::{fs, path::Path, env};
+
+use clap::{command, Parser};
+use slog::{Logger, Level, Drain, o};
+
+pub fn create_folder(root_path: &str) {
+    let path = Path::new(root_path);
+
+    if !path.exists() || !path.is_dir() {
+        // Create the folder if it doesn't exist
+        if let Err(err) = fs::create_dir(path) {
+            eprintln!("Error creating folder: {}", err);
+            // Handle the error as needed
+        } else {
+            println!("Folder created: {}", root_path);
+        }
+    } else {
+        println!("Folder already exists: {}", root_path);
+    }
+}
+
+#[derive(Parser, Clone)]
+#[command(author, version, about, long_about = None)]
+pub struct Cli {
+    /// Set the log level
+    #[arg(short, long)]
+    pub level: Option<String>,
+}
+
+pub fn setup_logger(cli: &Cli) -> Logger {
+    let log_level = if cli.level.is_some() {
+        let level = cli.level.as_ref().unwrap();
+        match level.as_ref() {
+            "trace" => Level::Trace,
+            "debug" => Level::Debug,
+            "info" => Level::Info,
+            "warn" => Level::Warning,
+            "error" => Level::Error,
+            _ => Level::Info,
+        }
+    } else {
+        let rust_log = env::var("RUST_LOG").unwrap_or_else(|_| String::from(""));
+        match rust_log.to_lowercase().as_str() {
+            "trace" => Level::Trace,
+            "debug" => Level::Debug,
+            "info" => Level::Info,
+            "warn" => Level::Warning,
+            "error" => Level::Error,
+            _ => Level::Info,
+        }
+    };
+
+    let decorator = slog_term::TermDecorator::new().build();
+    let drain = slog_term::CompactFormat::new(decorator).build().fuse();
+    let drain = slog_async::Async::new(drain).build().fuse();
+    let drain = drain.filter_level(log_level).fuse();
+    slog::Logger::root(drain, o!("version" => "0.5"))
+}
